Converted Converters to implicit conversions
